Login component added.

+++ src/xergor/modules/pages/settings/address/district/district.component.ts
@@ -1,15 +1,19 @@
-import { Component, OnDestroy, OnInit } from '@angular/core';
+import { AfterContentChecked, ChangeDetectorRef, Component, OnDestroy, OnInit } from '@angular/core';
 import { AbstractControl, FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { MatDialog } from '@angular/material/dialog';
+import { MatSnackBar } from '@angular/material/snack-bar';
 import { MatTableDataSource } from '@angular/material/table';
 import { ActivatedRoute, Router } from '@angular/router';
 import { catchError, first, throwError } from 'rxjs';
+import { ConfirmDialogComponent } from 'src/xergor/shared/components/confirm-dialog/confirm-dialog.component';
 import { Column } from 'src/xergor/shared/components/nsgrid/nsgrid.model';
 import { XHergorAlertService } from 'src/xergor/shared/components/xergor-alert';
 import { ICity } from 'src/xergor/shared/models/city.model';
 import { IDistrictData } from 'src/xergor/shared/models/district.model';
-import { DistrictService } from '../district.service';
-import { CityColumnCollection } from './districtDataColumn';
+import { XergorNavigationService } from 'src/xergor/shared/services/xergornavigation.service';
+import { DistrictService } from './district.service';
  
+import { CityColumnCollection } from './districtDataColumn';
 
 @Component({
   selector: 'app-district',
@@ -17,14 +21,12 @@ import { CityColumnCollection } from './districtDataColumn';
   styleUrls: ['./district.component.scss']
 })
 
-export class DistrictComponent implements OnInit, OnDestroy {
+export class DistrictComponent implements OnInit, OnDestroy  {
 
   keyword: string = 'cityName';
   f_cities!: MatTableDataSource<ICity[]>;
   column!: Array<Column>;
-
   g_keyword: string = 'cityName';
-
   f_districtForm!: FormGroup;
 
   district: IDistrictData = {
@@ -48,7 +50,10 @@ export class DistrictComponent implements OnInit, OnDestroy {
     private route: ActivatedRoute,
     private router: Router,
     private service: DistrictService,
-    private alertService: XHergorAlertService
+    private alertService: XHergorAlertService,
+    private navigationService: XergorNavigationService,
+    private dialog: MatDialog,
+    private snackBar: MatSnackBar
   ) {
     this.router.routeReuseStrategy.shouldReuseRoute = () => false;
 
@@ -72,7 +77,7 @@ export class DistrictComponent implements OnInit, OnDestroy {
         this.getDistrictById(id);
       }
       else {
-        this.router.navigate(['/account/address/district']);
+        this.router.navigate(['/settings/districts']);
       }
     });
   }
@@ -99,11 +104,11 @@ export class DistrictComponent implements OnInit, OnDestroy {
     if (this.submitted) {
       if (this.f_isAddMode) {
         this.createDistrict();
-        this.alertService.success('Kayıt başarılı.');
+        this.snackBar.open('Kayıt başarılı.', '', { duration: 2000, });
       }
       else {
         this.updateDistrict(this.district!.id);
-        this.alertService.success('Kayıt Güncellendi');
+        this.snackBar.open('Kayıt Güncellendi.', '', { duration: 2000, });
       }
     }
   }
@@ -114,43 +119,32 @@ export class DistrictComponent implements OnInit, OnDestroy {
     this.district!.districtCode = this.DistrictForm['districtCode'].value;
     this.district!.districtName = this.DistrictForm['districtName'].value;
     this.district!.inUse = this.DistrictForm['inUse'].value;
-
     if (this.district && this.district!.cityId === 0) {
       this.submitted = false;
       this.alertService.warn('Şehir zorunludur.');
       return;
     }
-
     if (this.district && this.district!.districtCode === '') {
       this.submitted = false;
       this.alertService.warn('İlçe kodu zorunludur.');
       return;
     }
-
     if (this.district && this.district!.districtName === '') {
       this.submitted = false;
       this.alertService.warn('İlçe adı zorunludur.');
       return;
     }
-
-
   }
 
   private createDistrict() {
-
     this.service.setDistrict(this.district)
       .pipe(first())
       .subscribe(
         (response: any) => {
-          this.alertService.success('User added successfully', { keepAfterRouteChange: true });
-          // this.router.navigate(['/account/address/district/'+this.district.id, { relativeTo: this.route }]); 
           this.district = response.data;
-          this.router.navigate(['/account/address/district/detail/' + this.district!.id]);
-
+          this.router.navigate(['/settings/districts/' + this.district!.id]);
           this.alertService.success(response);
-          console.log(response.data);
           this.loading = false;
-          this.getPulseClass();
         },
         error => {
           this.alertService.error(error);
@@ -162,26 +156,21 @@ export class DistrictComponent implements OnInit, OnDestroy {
    * @param id 
    */
   private updateDistrict(id: number) {
-
     this.district!.id = this.DistrictForm['id'].value;
-    // this.district!.cityId = this.DistrictForm['cityId'].value;
     this.district!.districtCode = this.DistrictForm['districtCode'].value;
     this.district!.districtName = this.DistrictForm['districtName'].value;
     this.district!.inUse = this.DistrictForm['inUse'].value;
-
     this.service.updateDistrict(id, this.district!)
       .pipe(catchError((err) => this.handleError(err)))
       .subscribe((response: any) => {
-        console.log('district updated: ' + response.data);
         this.district = response.data;
         this.loading = false;
-        this.router.navigate(['/account/address/district/detail/' + this.district!.id]);
+        this.router.navigate(['/settings/districts/' + this.district!.id]);
       },
         error => {
           this.alertService.error(error);
           this.loading = false;
         });
-
   }
 
   loadData() {
@@ -191,8 +180,6 @@ export class DistrictComponent implements OnInit, OnDestroy {
   getCities() {
     this.service.getCities().subscribe(l_cities => {
       this.f_cities = new MatTableDataSource(l_cities);
-      // let l_allCities = { id: 0, cityCode: '', cityName: '', inUse: false };
-      // this.f_cities.unshift(l_allCities);
     });
   }
 
@@ -208,7 +195,6 @@ export class DistrictComponent implements OnInit, OnDestroy {
           this.DistrictForm["inUse"].setValue(this.district?.inUse);
           this.DistrictForm["cityId"].setValue(this.district?.cityId);
           this.getCityById(this.district?.cityId);
-
         }
       });
   }
@@ -225,17 +211,35 @@ export class DistrictComponent implements OnInit, OnDestroy {
       });
   }
 
-
   delete(id: number) {
-    this.service.deleteDistrict(id)
-      .pipe(catchError((err: any) => this.handleError(err)))
-      .subscribe((res: any) => {
-        this.router.navigate(['/account/address/district']);
-      },
-        error => {
-          this.alertService.error(error);
-          this.loading = false;
-        });
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      data: {
+        message: 'Silmek istediğinizden emin misiniz?',
+        buttonText: {
+          ok: 'Evet',
+          cancel: 'Hayır'
+        }
+      }
+    });
+
+    dialogRef.afterClosed().subscribe((confirmed: boolean) => {
+      if (confirmed) {
+        this.service.deleteDistrict(id)
+          .pipe(catchError((err: any) => this.handleError(err)))
+          .subscribe((res: any) => {
+            this.router.navigate(['/settings/districts']);
+          },
+            error => {
+              this.alertService.error(error);
+              this.loading = false;
+            });
+
+        const a = document.createElement('a');
+        a.click();
+        a.remove();
+
+      }
+    });
   }
 
   /**
@@ -279,14 +283,15 @@ export class DistrictComponent implements OnInit, OnDestroy {
     this.DistrictForm["cityId"].markAsDirty();
   }
 
-  getPulseClass(): string {
-    let pulseClass = '';
-    if (this.f_districtForm.valid && this.f_districtForm.dirty) {
-      pulseClass = 'pulse';
-    }
-    else {
-      pulseClass = '';
-    }
-    return pulseClass;
+  back(): void {
+    this.navigationService.backToOldUrl();
+  }
+
+  public refreshContent() {
+    let currentUrl = this.router.url;
+    this.router.navigateByUrl('/', { skipLocationChange: true }).then(() => {
+      this.router.navigate([currentUrl]);
+    });
+
   }
 }
